discord_bot/requirements.txt
discord.py
flask
minio
